refactor: remove HttpRpc
