+++ src/main/java/com/limechain/rpc/ws/server/AppBean.java
@@ -1,4 +1,4 @@
-package com.limechain.rpc.http.server;
+package com.limechain.rpc.ws.server;
 
 import org.springframework.beans.BeansException;
 import org.springframework.context.ApplicationContext;
