+++ src/main/java/com/limechain/sync/warpsync/SyncedState.java
@@ -7,7 +7,7 @@ import com.limechain.network.protocol.blockannounce.NodeRole;
 import com.limechain.network.protocol.blockannounce.scale.BlockAnnounceHandshake;
 import com.limechain.network.protocol.grandpa.messages.commit.CommitMessage;
 import com.limechain.network.protocol.grandpa.messages.neighbour.NeighbourMessage;
-import com.limechain.rpc.http.server.AppBean;
+import com.limechain.rpc.ws.server.AppBean;
 import com.limechain.runtime.Runtime;
 import com.limechain.sync.JustificationVerifier;
 import io.emeraldpay.polkaj.types.Hash256;
