+++ src/test/java/com/limechain/lightclient/LightClientTest.java
@@ -1,7 +1,6 @@
 package com.limechain.lightclient;
 
 import com.limechain.network.Network;
-import com.limechain.rpc.http.server.HttpRpc;
 import com.limechain.rpc.ws.server.WebSocketRPC;
 import com.limechain.sync.warpsync.WarpSyncMachine;
 import org.junit.jupiter.api.BeforeEach;
@@ -13,10 +12,9 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-public class LightClientTest {
+class LightClientTest {
     private LightClient lightClient;
 
-    private HttpRpc httpRpc;
     private WebSocketRPC wsRpc;
     private String[] args;
 
@@ -31,11 +29,10 @@ public class LightClientTest {
 
     @BeforeEach
     public void setup() {
-        httpRpc = mock(HttpRpc.class);
         wsRpc = mock(WebSocketRPC.class);
         args = new String[]{"some args"};
 
-        lightClient = new LightClient(args, httpRpc, wsRpc);
+        lightClient = new LightClient(args, wsRpc);
     }
 
     @Test
@@ -48,7 +45,6 @@ public class LightClientTest {
 
         lightClient.stop();
 
-        verify(httpRpc, times(1)).stop();
         verify(wsRpc, times(1)).stop();
 
     }
